@@ -5,7 +5,7 @@ description = ""
 authors = ["Micah Lee <micah.lee@firstlook.media>"]
 
 [tool.poetry.dependencies]
-python = "^3.6.2"
+python = "^3.8"
 boto3 = "^1.18.28"
 click = "^8.0.1"
 
